# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "EnqueueJobBody": {
        "properties": {
          "job_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Job Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "job_type"
        ],
        "title": "EnqueueJobBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplacePlayerSpotlightsBody": {
        "description": "Full replace: existing rows for this game are deleted, then these are inserted.",
        "properties": {
          "spotlights": {
            "items": {
              "$ref": "#/components/schemas/PlayerSpotlightItem"
            },
            "title": "Spotlights",
            "type": "array"
          }
        },
        "title": "ReplacePlayerSpotlightsBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "AI-powered sports prediction API",
    "title": "Octobet API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/data-telemetry/detail": {
      "get": {
        "description": "Full provider-error samples + freshness/coverage for ops (audit #13).",
        "operationId": "data_telemetry_detail_internal_data_telemetry_detail_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Data Telemetry Detail",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/email-digest/run": {
      "post": {
        "description": "Send daily picks email when FEATURE_EMAIL_DIGEST=true and SMTP configured.",
        "operationId": "run_email_digest_cron_internal_email_digest_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Email Digest Cron",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/forecast-ledger/verify": {
      "get": {
        "description": "Recompute content-hash chain for the append-only forecast ledger.",
        "operationId": "verify_forecast_ledger_cron_internal_forecast_ledger_verify_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Forecast Ledger Cron",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/games/{game_id}/player-spotlights": {
      "delete": {
        "description": "Remove all spotlight rows for a game.",
        "operationId": "clear_game_player_spotlights_internal_games__game_id__player_spotlights_delete",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Game Player Spotlights",
        "tags": [
          "internal"
        ]
      },
      "put": {
        "description": "Replace all performer spotlight rows for a game (ETL / admin). Same auth as cron:\nX-Cron-Secret: <PUSH_CRON_SECRET>",
        "operationId": "replace_game_player_spotlights_internal_games__game_id__player_spotlights_put",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplacePlayerSpotlightsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Game Player Spotlights",
        "tags": [
          "internal"
        ]
      }
    },
...
    "/internal/historical-backfill/run": {
      "post": {
        "description": "M-07: Sync 2+ seasons per league from ClearSports (CLEARSPORTS_API_KEY).\nIdempotent upserts; returns decisive finished-game counts per league.\nSchedule weekly or run once after deploy. Optionally chain predictions for backtest labels.",
        "operationId": "historical_backfill_run_internal_historical_backfill_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalBackfillBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Historical Backfill Run",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/jobs/enqueue": {
      "post": {
        "description": "Enqueue background work (Redis list; Imp #52).",
        "operationId": "enqueue_internal_job_internal_jobs_enqueue_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueJobBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue Internal Job",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/jobs/process-one": {
      "post": {
        "description": "Dequeue and return one job for worker processing.",
        "operationId": "process_one_job_internal_jobs_process_one_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cron-Secret",
            "required": false,
            "schema": {
              "title": "X-Cron-Secret",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Process One Job",
        "tags": [
          "internal"
        ]
      }
    },
...
  }
}